@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.10.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
@@ -28,20 +28,22 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.3
-  firebase_core: ^1.1.1
-  firebase_auth: ^1.1.4
-  google_sign_in: ^5.0.3
-  cloud_functions: ^1.0.7
-  cloud_firestore: ^2.1.0
+  firebase_core: ^1.3.0
+  firebase_auth: ^2.0.0
+  google_sign_in: ^5.0.4
+  cloud_functions: ^2.0.0
+  cloud_firestore: ^2.3.0
   firebase_core_platform_interface: ^4.0.1
   flutter_spinkit: ^5.0.0
-  flutter_zoom_drawer: ^2.0.2+1
+  flutter_zoom_drawer: ^2.0.2+4
   intl: ^0.17.0
   sliding_up_panel: ^2.0.0+1
   flutter_riverpod: ^0.14.0+3
-  shared_preferences: ^2.0.5
-  file_picker:
-  firebase_storage: ^8.0.6
+  shared_preferences: ^2.0.6
+  file_picker: ^3.0.3
+  firebase_storage: ^9.0.0
+  agora_rtc_engine: ^4.0.2
+  permission_handler: ^8.1.2
 
 
 
