@@ -3,12 +3,12 @@ import 'package:flutter/material.dart';
 
 import 'show_alert_dialog.dart';
 
-Future<bool> showExceptionAlertDialog(BuildContext context,
-        {String title, Exception exception}) =>
+Future<bool?> showExceptionAlertDialog(BuildContext context,
+        {String? title, Exception? exception}) =>
     showAlertDialog(context,
         title: title, defaultActionText: 'OK', content: _message(exception));
 
-String _message(Exception exception) {
+String? _message(Exception? exception) {
   if (exception is FirebaseAuthException) {
     return exception.message;
   }
